+++ lib/defaults/src/pragma_map.cpp
@@ -16,7 +16,7 @@
 
 #include <metashell/defaults/pragma_map.hpp>
 
-#include <metashell/pragma/config.hpp>
+#include <metashell/pragma/all_config.hpp>
 #include <metashell/pragma/config_load.hpp>
 #include <metashell/pragma/config_show.hpp>
 #include <metashell/pragma/echo.hpp>
@@ -173,7 +173,8 @@ namespace metashell
 
       result.emplace(sv{"ls"}, std::make_unique<pragma::ls>());
 
-      result.emplace(sv{"config"}, std::make_unique<pragma::config>());
+      result.emplace(
+          sv{"all", "config"}, std::make_unique<pragma::all_config>());
 
       result.emplace(
           sv{"config", "show"}, std::make_unique<pragma::config_show>());